mutinamespace in python
