+++ package.json
@@ -11,7 +11,7 @@
     "@typespec/http-server-javascript": "^0.58.0-alpha.6",
     "@typespec/http-client-java": "^0.1.5",
     "@azure-tools/typespec-ts": "^0.36.0",
-    "@typespec/http-client-python": "^0.4.4",
+    "@typespec/http-client-python": "https://artprodcus3.artifacts.visualstudio.com/A0fb41ef4-5012-48a9-bf39-4ee3de03ee35/29ec6040-b234-4e31-b139-33dc4287b756/_apis/artifact/cGlwZWxpbmVhcnRpZmFjdDovL2F6dXJlLXNkay9wcm9qZWN0SWQvMjllYzYwNDAtYjIzNC00ZTMxLWIxMzktMzNkYzQyODdiNzU2L2J1aWxkSWQvNDQzMTQ0OC9hcnRpZmFjdE5hbWUvYnVpbGRfYXJ0aWZhY3RzX3B5dGhvbg2/content?format=file&subPath=%2Fpackages%2Ftypespec-http-client-python-0.4.4.tgz",
     "@typespec/json-schema": "^0.63.0",
     "@typespec/openapi": "^0.63.0",
     "@typespec/openapi3": "^0.63.0",
