+++ petstore/clients/python/sample/sample.py
