+++ todoApp/clients/python/todo/models/_models.py
@@ -345,90 +345,6 @@ class GetResponse(_model_base.Model):
         super().__init__(*args, **kwargs)
 
 
-class InvalidTodoItem(ApiError):
-    """InvalidTodoItem.
-
-    All required parameters must be populated in order to send to server.
-
-    :ivar code: A machine readable error code. Required.
-    :vartype code: str
-    :ivar message: A human readable message. Required.
-    :vartype message: str
-    """
-
-    @overload
-    def __init__(
-        self,
-        *,
-        code: str,
-        message: str,
-    ) -> None: ...
-
-    @overload
-    def __init__(self, mapping: Mapping[str, Any]) -> None:
-        """
-        :param mapping: raw JSON to initialize the model.
-        :type mapping: Mapping[str, Any]
-        """
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-
-
-class InvalidUserResponse(ApiError):
-    """The user is invalid (e.g. forgot to enter email address).
-
-    Readonly variables are only populated by the server, and will be ignored when sending a request.
-
-    All required parameters must be populated in order to send to server.
-
-    :ivar message: A human readable message. Required.
-    :vartype message: str
-    :ivar code: Required. Default value is "invalid-user".
-    :vartype code: str
-    """
-
-    code: Literal["invalid-user"] = rest_field()
-    """Required. Default value is \"invalid-user\"."""
-
-    @overload
-    def __init__(
-        self,
-        *,
-        message: str,
-    ) -> None: ...
-
-    @overload
-    def __init__(self, mapping: Mapping[str, Any]) -> None:
-        """
-        :param mapping: raw JSON to initialize the model.
-        :type mapping: Mapping[str, Any]
-        """
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-        self.code: Literal["invalid-user"] = "invalid-user"
-
-
-class NotFoundErrorResponse(_model_base.Model):
-    """NotFoundErrorResponse.
-
-    Readonly variables are only populated by the server, and will be ignored when sending a request.
-
-    All required parameters must be populated in order to send to server.
-
-    :ivar code: Required. Default value is "not-found".
-    :vartype code: str
-    """
-
-    code: Literal["not-found"] = rest_field()
-    """Required. Default value is \"not-found\"."""
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-        self.code: Literal["not-found"] = "not-found"
-
-
 class Standard4XXResponse(ApiError):
     """Something is wrong with you.
 
@@ -640,51 +556,6 @@ class ToDoItemMultipartRequest(_model_base.Model):
         super().__init__(*args, **kwargs)
 
 
-class TodoItemPatch(_model_base.Model):
-    """TodoItemPatch.
-
-    :ivar title: The item's title.
-    :vartype title: str
-    :ivar assigned_to: User that the todo is assigned to.
-    :vartype assigned_to: int
-    :ivar description: A longer description of the todo item in markdown format.
-    :vartype description: str
-    :ivar status: The status of the todo item. Is one of the following types:
-     Literal["NotStarted"], Literal["InProgress"], Literal["Completed"]
-    :vartype status: str or str or str
-    """
-
-    title: Optional[str] = rest_field()
-    """The item's title."""
-    assigned_to: Optional[int] = rest_field(name="assignedTo")
-    """User that the todo is assigned to."""
-    description: Optional[str] = rest_field()
-    """A longer description of the todo item in markdown format."""
-    status: Optional[Literal["NotStarted", "InProgress", "Completed"]] = rest_field()
-    """The status of the todo item. Is one of the following types: Literal[\"NotStarted\"],
-     Literal[\"InProgress\"], Literal[\"Completed\"]"""
-
-    @overload
-    def __init__(
-        self,
-        *,
-        title: Optional[str] = None,
-        assigned_to: Optional[int] = None,
-        description: Optional[str] = None,
-        status: Optional[Literal["NotStarted", "InProgress", "Completed"]] = None,
-    ) -> None: ...
-
-    @overload
-    def __init__(self, mapping: Mapping[str, Any]) -> None:
-        """
-        :param mapping: raw JSON to initialize the model.
-        :type mapping: Mapping[str, Any]
-        """
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-
-
 class TodoLabelRecord(_model_base.Model):
     """TodoLabelRecord.
 
@@ -838,38 +709,3 @@ class User(_model_base.Model):
 
     def __init__(self, *args: Any, **kwargs: Any) -> None:
         super().__init__(*args, **kwargs)
-
-
-class UserExistsResponse(ApiError):
-    """The user already exists.
-
-    Readonly variables are only populated by the server, and will be ignored when sending a request.
-
-    All required parameters must be populated in order to send to server.
-
-    :ivar message: A human readable message. Required.
-    :vartype message: str
-    :ivar code: Required. Default value is "user-exists".
-    :vartype code: str
-    """
-
-    code: Literal["user-exists"] = rest_field()
-    """Required. Default value is \"user-exists\"."""
-
-    @overload
-    def __init__(
-        self,
-        *,
-        message: str,
-    ) -> None: ...
-
-    @overload
-    def __init__(self, mapping: Mapping[str, Any]) -> None:
-        """
-        :param mapping: raw JSON to initialize the model.
-        :type mapping: Mapping[str, Any]
-        """
-
-    def __init__(self, *args: Any, **kwargs: Any) -> None:
-        super().__init__(*args, **kwargs)
-        self.code: Literal["user-exists"] = "user-exists"
