+++ todoApp/clients/python/todo/todoitems/aio/operations/_operations.py
@@ -1,4 +1,3 @@
-# pylint: disable=too-many-lines
 # coding=utf-8
 
 from io import IOBase
@@ -21,20 +20,18 @@ from corehttp.rest import AsyncHttpResponse, HttpRequest
 from corehttp.runtime.pipeline import PipelineResponse
 from corehttp.utils import case_insensitive_dict
 
-from ... import _model_base, models as _models
-from ..._model_base import SdkJSONEncoder, _deserialize, _failsafe_deserialize
-from ..._vendor import prepare_multipart_form_data
+from ... import models as _models2
+from .... import _model_base, models as _models3
+from ...._model_base import SdkJSONEncoder, _deserialize, _failsafe_deserialize
+from ...._vendor import prepare_multipart_form_data
+from ...attachments.aio.operations._operations import TodoItemsAttachmentsOperations
 from ...operations._operations import (
-    build_todo_items_attachments_create_file_attachment_request,
-    build_todo_items_attachments_create_json_attachment_request,
-    build_todo_items_attachments_list_request,
     build_todo_items_create_form_request,
     build_todo_items_create_json_request,
     build_todo_items_delete_request,
     build_todo_items_get_request,
     build_todo_items_list_request,
     build_todo_items_update_request,
-    build_users_create_request,
 )
 
 if sys.version_info >= (3, 9):
@@ -42,147 +39,9 @@ if sys.version_info >= (3, 9):
 else:
     from typing import MutableMapping  # type: ignore
 JSON = MutableMapping[str, Any]  # pylint: disable=unsubscriptable-object
+_Unset: Any = object()
 T = TypeVar("T")
 ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T, Dict[str, Any]], Any]]
-_Unset: Any = object()
-
-
-class UsersOperations:
-    """
-    .. warning::
-        **DO NOT** instantiate this class directly.
-
-        Instead, you should access the following operations through
-        :class:`~todo.aio.TodoClient`'s
-        :attr:`users` attribute.
-    """
-
-    def __init__(self, *args, **kwargs) -> None:
-        input_args = list(args)
-        self._client = input_args.pop(0) if input_args else kwargs.pop("client")
-        self._config = input_args.pop(0) if input_args else kwargs.pop("config")
-        self._serialize = input_args.pop(0) if input_args else kwargs.pop("serializer")
-        self._deserialize = input_args.pop(0) if input_args else kwargs.pop("deserializer")
-
-    @overload
-    async def create(
-        self, user: _models.User, *, content_type: str = "application/json", **kwargs: Any
-    ) -> _models.CreateResponse:
-        """create.
-
-        :param user: Required.
-        :type user: ~todo.models.User
-        :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
-         Default value is "application/json".
-        :paramtype content_type: str
-        :return: CreateResponse. The CreateResponse is compatible with MutableMapping
-        :rtype: ~todo.models.CreateResponse
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-
-    @overload
-    async def create(
-        self, user: JSON, *, content_type: str = "application/json", **kwargs: Any
-    ) -> _models.CreateResponse:
-        """create.
-
-        :param user: Required.
-        :type user: JSON
-        :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
-         Default value is "application/json".
-        :paramtype content_type: str
-        :return: CreateResponse. The CreateResponse is compatible with MutableMapping
-        :rtype: ~todo.models.CreateResponse
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-
-    @overload
-    async def create(
-        self, user: IO[bytes], *, content_type: str = "application/json", **kwargs: Any
-    ) -> _models.CreateResponse:
-        """create.
-
-        :param user: Required.
-        :type user: IO[bytes]
-        :keyword content_type: Body Parameter content-type. Content type parameter for binary body.
-         Default value is "application/json".
-        :paramtype content_type: str
-        :return: CreateResponse. The CreateResponse is compatible with MutableMapping
-        :rtype: ~todo.models.CreateResponse
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-
-    async def create(self, user: Union[_models.User, JSON, IO[bytes]], **kwargs: Any) -> _models.CreateResponse:
-        """create.
-
-        :param user: Is one of the following types: User, JSON, IO[bytes] Required.
-        :type user: ~todo.models.User or JSON or IO[bytes]
-        :return: CreateResponse. The CreateResponse is compatible with MutableMapping
-        :rtype: ~todo.models.CreateResponse
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-        error_map: MutableMapping = {
-            304: ResourceNotModifiedError,
-        }
-        error_map.update(kwargs.pop("error_map", {}) or {})
-
-        _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
-        _params = kwargs.pop("params", {}) or {}
-
-        content_type: Optional[str] = kwargs.pop("content_type", _headers.pop("Content-Type", None))
-        cls: ClsType[_models.CreateResponse] = kwargs.pop("cls", None)
-
-        content_type = content_type or "application/json"
-        _content = None
-        if isinstance(user, (IOBase, bytes)):
-            _content = user
-        else:
-            _content = json.dumps(user, cls=SdkJSONEncoder, exclude_readonly=True)  # type: ignore
-
-        _request = build_users_create_request(
-            content_type=content_type,
-            content=_content,
-            headers=_headers,
-            params=_params,
-        )
-        path_format_arguments = {
-            "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
-        }
-        _request.url = self._client.format_url(_request.url, **path_format_arguments)
-
-        _stream = kwargs.pop("stream", False)
-        pipeline_response: PipelineResponse = await self._client.pipeline.run(_request, stream=_stream, **kwargs)
-
-        response = pipeline_response.http_response
-
-        if response.status_code not in [200]:
-            if _stream:
-                try:
-                    await response.read()  # Load the body in memory and close the socket
-                except (StreamConsumedError, StreamClosedError):
-                    pass
-            map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = None
-            if response.status_code == 409:
-                error = _failsafe_deserialize(_models.UserExistsResponse, response.json())
-                raise ResourceExistsError(response=response, model=error)
-            elif response.status_code == 422:
-                error = _failsafe_deserialize(_models.InvalidUserResponse, response.json())
-            elif 400 <= response.status_code <= 499:
-                error = _failsafe_deserialize(_models.Standard4XXResponse, response.json())
-            elif 500 <= response.status_code <= 599:
-                error = _failsafe_deserialize(_models.Standard5XXResponse, response.json())
-            raise HttpResponseError(response=response, model=error)
-
-        if _stream:
-            deserialized = response.iter_bytes()
-        else:
-            deserialized = _deserialize(_models.CreateResponse, response.json())
-
-        if cls:
-            return cls(pipeline_response, deserialized, {})  # type: ignore
-
-        return deserialized  # type: ignore
 
 
 class TodoItemsOperations:
@@ -208,7 +67,7 @@ class TodoItemsOperations:
 
     def list(
         self, *, limit: Optional[int] = None, offset: Optional[int] = None, **kwargs: Any
-    ) -> AsyncIterable["_models.TodoItem"]:
+    ) -> AsyncIterable["_models3.TodoItem"]:
         """list.
 
         :keyword limit: The limit to the number of items. Default value is None.
@@ -222,7 +81,7 @@ class TodoItemsOperations:
         _headers = kwargs.pop("headers", {}) or {}
         _params = kwargs.pop("params", {}) or {}
 
-        cls: ClsType[List[_models.TodoItem]] = kwargs.pop("cls", None)
+        cls: ClsType[List[_models3.TodoItem]] = kwargs.pop("cls", None)
 
         error_map: MutableMapping = {
             304: ResourceNotModifiedError,
@@ -258,7 +117,7 @@ class TodoItemsOperations:
 
         async def extract_data(pipeline_response):
             deserialized = pipeline_response.http_response.json()
-            list_of_elem = _deserialize(List[_models.TodoItem], deserialized["items"])
+            list_of_elem = _deserialize(List[_models3.TodoItem], deserialized["items"])
             if cls:
                 list_of_elem = cls(list_of_elem)  # type: ignore
             return deserialized.get("nextLink") or None, AsyncList(list_of_elem)
@@ -274,9 +133,9 @@ class TodoItemsOperations:
                 map_error(status_code=response.status_code, response=response, error_map=error_map)
                 error = None
                 if 400 <= response.status_code <= 499:
-                    error = _failsafe_deserialize(_models.Standard4XXResponse, response.json())
+                    error = _failsafe_deserialize(_models3.Standard4XXResponse, response.json())
                 elif 500 <= response.status_code <= 599:
-                    error = _failsafe_deserialize(_models.Standard5XXResponse, response.json())
+                    error = _failsafe_deserialize(_models3.Standard5XXResponse, response.json())
                 raise HttpResponseError(response=response, model=error)
 
             return pipeline_response
@@ -286,7 +145,7 @@ class TodoItemsOperations:
     @overload
     async def create_json(
         self, body: JSON, *, content_type: str = "application/json", **kwargs: Any
-    ) -> _models.CreateJsonResponse:
+    ) -> _models3.CreateJsonResponse:
         """create_json.
 
         :param body: Required.
@@ -303,11 +162,11 @@ class TodoItemsOperations:
     async def create_json(
         self,
         *,
-        item: _models.TodoItem,
+        item: _models3.TodoItem,
         content_type: str = "application/json",
-        attachments: Optional[List[_models.TodoAttachment]] = None,
+        attachments: Optional[List[_models3.TodoAttachment]] = None,
         **kwargs: Any
-    ) -> _models.CreateJsonResponse:
+    ) -> _models3.CreateJsonResponse:
         """create_json.
 
         :keyword item: Required.
@@ -325,7 +184,7 @@ class TodoItemsOperations:
     @overload
     async def create_json(
         self, body: IO[bytes], *, content_type: str = "application/json", **kwargs: Any
-    ) -> _models.CreateJsonResponse:
+    ) -> _models3.CreateJsonResponse:
         """create_json.
 
         :param body: Required.
@@ -342,10 +201,10 @@ class TodoItemsOperations:
         self,
         body: Union[JSON, IO[bytes]] = _Unset,
         *,
-        item: _models.TodoItem = _Unset,
-        attachments: Optional[List[_models.TodoAttachment]] = None,
+        item: _models3.TodoItem = _Unset,
+        attachments: Optional[List[_models3.TodoAttachment]] = None,
         **kwargs: Any
-    ) -> _models.CreateJsonResponse:
+    ) -> _models3.CreateJsonResponse:
         """create_json.
 
         :param body: Is either a JSON type or a IO[bytes] type. Required.
@@ -367,7 +226,7 @@ class TodoItemsOperations:
         _params = kwargs.pop("params", {}) or {}
 
         content_type: Optional[str] = kwargs.pop("content_type", _headers.pop("content-type", None))
-        cls: ClsType[_models.CreateJsonResponse] = kwargs.pop("cls", None)
+        cls: ClsType[_models3.CreateJsonResponse] = kwargs.pop("cls", None)
 
         if body is _Unset:
             if item is _Unset:
@@ -406,17 +265,17 @@ class TodoItemsOperations:
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             error = None
             if response.status_code == 422:
-                error = _failsafe_deserialize(_models.InvalidTodoItem, response.json())
+                error = _failsafe_deserialize(_models2.InvalidTodoItem, response.json())
             elif 400 <= response.status_code <= 499:
-                error = _failsafe_deserialize(_models.Standard4XXResponse, response.json())
+                error = _failsafe_deserialize(_models3.Standard4XXResponse, response.json())
             elif 500 <= response.status_code <= 599:
-                error = _failsafe_deserialize(_models.Standard5XXResponse, response.json())
+                error = _failsafe_deserialize(_models3.Standard5XXResponse, response.json())
             raise HttpResponseError(response=response, model=error)
 
         if _stream:
             deserialized = response.iter_bytes()
         else:
-            deserialized = _deserialize(_models.CreateJsonResponse, response.json())
+            deserialized = _deserialize(_models3.CreateJsonResponse, response.json())
 
         if cls:
             return cls(pipeline_response, deserialized, {})  # type: ignore
@@ -424,7 +283,7 @@ class TodoItemsOperations:
         return deserialized  # type: ignore
 
     @overload
-    async def create_form(self, body: _models.ToDoItemMultipartRequest, **kwargs: Any) -> _models.CreateFormResponse:
+    async def create_form(self, body: _models3.ToDoItemMultipartRequest, **kwargs: Any) -> _models3.CreateFormResponse:
         """create_form.
 
         :param body: Required.
@@ -435,7 +294,7 @@ class TodoItemsOperations:
         """
 
     @overload
-    async def create_form(self, body: JSON, **kwargs: Any) -> _models.CreateFormResponse:
+    async def create_form(self, body: JSON, **kwargs: Any) -> _models3.CreateFormResponse:
         """create_form.
 
         :param body: Required.
@@ -446,8 +305,8 @@ class TodoItemsOperations:
         """
 
     async def create_form(
-        self, body: Union[_models.ToDoItemMultipartRequest, JSON], **kwargs: Any
-    ) -> _models.CreateFormResponse:
+        self, body: Union[_models3.ToDoItemMultipartRequest, JSON], **kwargs: Any
+    ) -> _models3.CreateFormResponse:
         """create_form.
 
         :param body: Is either a ToDoItemMultipartRequest type or a JSON type. Required.
@@ -464,7 +323,7 @@ class TodoItemsOperations:
         _headers = kwargs.pop("headers", {}) or {}
         _params = kwargs.pop("params", {}) or {}
 
-        cls: ClsType[_models.CreateFormResponse] = kwargs.pop("cls", None)
+        cls: ClsType[_models3.CreateFormResponse] = kwargs.pop("cls", None)
 
         _body = body.as_dict() if isinstance(body, _model_base.Model) else body
         _file_fields: List[str] = ["attachments"]
@@ -496,24 +355,24 @@ class TodoItemsOperations:
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             error = None
             if response.status_code == 422:
-                error = _failsafe_deserialize(_models.InvalidTodoItem, response.json())
+                error = _failsafe_deserialize(_models2.InvalidTodoItem, response.json())
             elif 400 <= response.status_code <= 499:
-                error = _failsafe_deserialize(_models.Standard4XXResponse, response.json())
+                error = _failsafe_deserialize(_models3.Standard4XXResponse, response.json())
             elif 500 <= response.status_code <= 599:
-                error = _failsafe_deserialize(_models.Standard5XXResponse, response.json())
+                error = _failsafe_deserialize(_models3.Standard5XXResponse, response.json())
             raise HttpResponseError(response=response, model=error)
 
         if _stream:
             deserialized = response.iter_bytes()
         else:
-            deserialized = _deserialize(_models.CreateFormResponse, response.json())
+            deserialized = _deserialize(_models3.CreateFormResponse, response.json())
 
         if cls:
             return cls(pipeline_response, deserialized, {})  # type: ignore
 
         return deserialized  # type: ignore
 
-    async def get(self, id: int, **kwargs: Any) -> _models.GetResponse:
+    async def get(self, id: int, **kwargs: Any) -> _models3.GetResponse:
         """get.
 
         :param id: Required.
@@ -532,7 +391,7 @@ class TodoItemsOperations:
         _headers = kwargs.pop("headers", {}) or {}
         _params = kwargs.pop("params", {}) or {}
 
-        cls: ClsType[_models.GetResponse] = kwargs.pop("cls", None)
+        cls: ClsType[_models3.GetResponse] = kwargs.pop("cls", None)
 
         _request = build_todo_items_get_request(
             id=id,
@@ -558,14 +417,14 @@ class TodoItemsOperations:
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             error = None
             if response.status_code == 404:
-                error = _failsafe_deserialize(_models.NotFoundErrorResponse, response.json())
+                error = _failsafe_deserialize(_models2.NotFoundErrorResponse, response.json())
                 raise ResourceNotFoundError(response=response, model=error)
             raise HttpResponseError(response=response, model=error)
 
         if _stream:
             deserialized = response.iter_bytes()
         else:
-            deserialized = _deserialize(_models.GetResponse, response.json())
+            deserialized = _deserialize(_models3.GetResponse, response.json())
 
         if cls:
             return cls(pipeline_response, deserialized, {})  # type: ignore
@@ -576,11 +435,11 @@ class TodoItemsOperations:
     async def update(
         self,
         id: int,
-        patch: _models.TodoItemPatch,
+        patch: _models2.TodoItemPatch,
         *,
         content_type: str = "application/merge-patch+json",
         **kwargs: Any
-    ) -> _models.UpdateResponse:
+    ) -> _models3.UpdateResponse:
         """update.
 
         :param id: Required.
@@ -598,7 +457,7 @@ class TodoItemsOperations:
     @overload
     async def update(
         self, id: int, patch: JSON, *, content_type: str = "application/merge-patch+json", **kwargs: Any
-    ) -> _models.UpdateResponse:
+    ) -> _models3.UpdateResponse:
         """update.
 
         :param id: Required.
@@ -616,7 +475,7 @@ class TodoItemsOperations:
     @overload
     async def update(
         self, id: int, patch: IO[bytes], *, content_type: str = "application/merge-patch+json", **kwargs: Any
-    ) -> _models.UpdateResponse:
+    ) -> _models3.UpdateResponse:
         """update.
 
         :param id: Required.
@@ -632,8 +491,8 @@ class TodoItemsOperations:
         """
 
     async def update(
-        self, id: int, patch: Union[_models.TodoItemPatch, JSON, IO[bytes]], **kwargs: Any
-    ) -> _models.UpdateResponse:
+        self, id: int, patch: Union[_models2.TodoItemPatch, JSON, IO[bytes]], **kwargs: Any
+    ) -> _models3.UpdateResponse:
         """update.
 
         :param id: Required.
@@ -656,7 +515,7 @@ class TodoItemsOperations:
         _params = kwargs.pop("params", {}) or {}
 
         content_type: Optional[str] = kwargs.pop("content_type", _headers.pop("content-type", None))
-        cls: ClsType[_models.UpdateResponse] = kwargs.pop("cls", None)
+        cls: ClsType[_models3.UpdateResponse] = kwargs.pop("cls", None)
 
         content_type = content_type or "application/merge-patch+json"
         _content = None
@@ -694,7 +553,7 @@ class TodoItemsOperations:
         if _stream:
             deserialized = response.iter_bytes()
         else:
-            deserialized = _deserialize(_models.UpdateResponse, response.json())
+            deserialized = _deserialize(_models3.UpdateResponse, response.json())
 
         if cls:
             return cls(pipeline_response, deserialized, {})  # type: ignore
@@ -739,310 +598,12 @@ class TodoItemsOperations:
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             error = None
             if response.status_code == 404:
-                error = _failsafe_deserialize(_models.NotFoundErrorResponse, response.json())
+                error = _failsafe_deserialize(_models2.NotFoundErrorResponse, response.json())
                 raise ResourceNotFoundError(response=response, model=error)
             elif 400 <= response.status_code <= 499:
-                error = _failsafe_deserialize(_models.Standard4XXResponse, response.json())
+                error = _failsafe_deserialize(_models3.Standard4XXResponse, response.json())
             elif 500 <= response.status_code <= 599:
-                error = _failsafe_deserialize(_models.Standard5XXResponse, response.json())
-            raise HttpResponseError(response=response, model=error)
-
-        if cls:
-            return cls(pipeline_response, None, {})  # type: ignore
-
-
-class TodoItemsAttachmentsOperations:
-    """
-    .. warning::
-        **DO NOT** instantiate this class directly.
-
-        Instead, you should access the following operations through
-        :class:`~todo.aio.TodoClient`'s
-        :attr:`attachments` attribute.
-    """
-
-    def __init__(self, *args, **kwargs) -> None:
-        input_args = list(args)
-        self._client = input_args.pop(0) if input_args else kwargs.pop("client")
-        self._config = input_args.pop(0) if input_args else kwargs.pop("config")
-        self._serialize = input_args.pop(0) if input_args else kwargs.pop("serializer")
-        self._deserialize = input_args.pop(0) if input_args else kwargs.pop("deserializer")
-
-    def list(self, item_id: int, **kwargs: Any) -> AsyncIterable["_models.TodoAttachment"]:
-        """list.
-
-        :param item_id: Required.
-        :type item_id: int
-        :return: An iterator like instance of TodoAttachment
-        :rtype: ~corehttp.paging.AsyncItemPaged[~todo.models.TodoAttachment]
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-        _headers = kwargs.pop("headers", {}) or {}
-        _params = kwargs.pop("params", {}) or {}
-
-        cls: ClsType[List[_models.TodoAttachment]] = kwargs.pop("cls", None)
-
-        error_map: MutableMapping = {
-            304: ResourceNotModifiedError,
-        }
-        error_map.update(kwargs.pop("error_map", {}) or {})
-
-        def prepare_request(next_link=None):
-            if not next_link:
-
-                _request = build_todo_items_attachments_list_request(
-                    item_id=item_id,
-                    headers=_headers,
-                    params=_params,
-                )
-                path_format_arguments = {
-                    "endpoint": self._serialize.url(
-                        "self._config.endpoint", self._config.endpoint, "str", skip_quote=True
-                    ),
-                }
-                _request.url = self._client.format_url(_request.url, **path_format_arguments)
-
-            else:
-                _request = HttpRequest("GET", next_link)
-                path_format_arguments = {
-                    "endpoint": self._serialize.url(
-                        "self._config.endpoint", self._config.endpoint, "str", skip_quote=True
-                    ),
-                }
-                _request.url = self._client.format_url(_request.url, **path_format_arguments)
-
-            return _request
-
-        async def extract_data(pipeline_response):
-            deserialized = pipeline_response.http_response.json()
-            list_of_elem = _deserialize(List[_models.TodoAttachment], deserialized["items"])
-            if cls:
-                list_of_elem = cls(list_of_elem)  # type: ignore
-            return None, AsyncList(list_of_elem)
-
-        async def get_next(next_link=None):
-            _request = prepare_request(next_link)
-
-            _stream = False
-            pipeline_response: PipelineResponse = await self._client.pipeline.run(_request, stream=_stream, **kwargs)
-            response = pipeline_response.http_response
-
-            if response.status_code not in [200]:
-                map_error(status_code=response.status_code, response=response, error_map=error_map)
-                error = None
-                if response.status_code == 404:
-                    error = _failsafe_deserialize(_models.NotFoundErrorResponse, response.json())
-                    raise ResourceNotFoundError(response=response, model=error)
-                elif 400 <= response.status_code <= 499:
-                    error = _failsafe_deserialize(_models.Standard4XXResponse, response.json())
-                elif 500 <= response.status_code <= 599:
-                    error = _failsafe_deserialize(_models.Standard5XXResponse, response.json())
-                raise HttpResponseError(response=response, model=error)
-
-            return pipeline_response
-
-        return AsyncItemPaged(get_next, extract_data)
-
-    @overload
-    async def create_json_attachment(
-        self, item_id: int, contents: _models.TodoAttachment, *, content_type: str = "application/json", **kwargs: Any
-    ) -> None:
-        """create_json_attachment.
-
-        :param item_id: Required.
-        :type item_id: int
-        :param contents: Required.
-        :type contents: ~todo.models.TodoAttachment
-        :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
-         Default value is "application/json".
-        :paramtype content_type: str
-        :return: None
-        :rtype: None
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-
-    @overload
-    async def create_json_attachment(
-        self, item_id: int, contents: JSON, *, content_type: str = "application/json", **kwargs: Any
-    ) -> None:
-        """create_json_attachment.
-
-        :param item_id: Required.
-        :type item_id: int
-        :param contents: Required.
-        :type contents: JSON
-        :keyword content_type: Body Parameter content-type. Content type parameter for JSON body.
-         Default value is "application/json".
-        :paramtype content_type: str
-        :return: None
-        :rtype: None
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-
-    @overload
-    async def create_json_attachment(
-        self, item_id: int, contents: IO[bytes], *, content_type: str = "application/json", **kwargs: Any
-    ) -> None:
-        """create_json_attachment.
-
-        :param item_id: Required.
-        :type item_id: int
-        :param contents: Required.
-        :type contents: IO[bytes]
-        :keyword content_type: Body Parameter content-type. Content type parameter for binary body.
-         Default value is "application/json".
-        :paramtype content_type: str
-        :return: None
-        :rtype: None
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-
-    async def create_json_attachment(
-        self, item_id: int, contents: Union[_models.TodoAttachment, JSON, IO[bytes]], **kwargs: Any
-    ) -> None:
-        """create_json_attachment.
-
-        :param item_id: Required.
-        :type item_id: int
-        :param contents: Is one of the following types: TodoAttachment, JSON, IO[bytes] Required.
-        :type contents: ~todo.models.TodoAttachment or JSON or IO[bytes]
-        :return: None
-        :rtype: None
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-        error_map: MutableMapping = {
-            304: ResourceNotModifiedError,
-        }
-        error_map.update(kwargs.pop("error_map", {}) or {})
-
-        _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
-        _params = kwargs.pop("params", {}) or {}
-
-        content_type: Optional[str] = kwargs.pop("content_type", _headers.pop("content-type", None))
-        cls: ClsType[None] = kwargs.pop("cls", None)
-
-        content_type = content_type or "application/json"
-        _content = None
-        if isinstance(contents, (IOBase, bytes)):
-            _content = contents
-        else:
-            _content = json.dumps(contents, cls=SdkJSONEncoder, exclude_readonly=True)  # type: ignore
-
-        _request = build_todo_items_attachments_create_json_attachment_request(
-            item_id=item_id,
-            content_type=content_type,
-            content=_content,
-            headers=_headers,
-            params=_params,
-        )
-        path_format_arguments = {
-            "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
-        }
-        _request.url = self._client.format_url(_request.url, **path_format_arguments)
-
-        _stream = False
-        pipeline_response: PipelineResponse = await self._client.pipeline.run(_request, stream=_stream, **kwargs)
-
-        response = pipeline_response.http_response
-
-        if response.status_code not in [204]:
-            map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = None
-            if response.status_code == 404:
-                error = _failsafe_deserialize(_models.NotFoundErrorResponse, response.json())
-                raise ResourceNotFoundError(response=response, model=error)
-            elif 400 <= response.status_code <= 499:
-                error = _failsafe_deserialize(_models.Standard4XXResponse, response.json())
-            elif 500 <= response.status_code <= 599:
-                error = _failsafe_deserialize(_models.Standard5XXResponse, response.json())
-            raise HttpResponseError(response=response, model=error)
-
-        if cls:
-            return cls(pipeline_response, None, {})  # type: ignore
-
-    @overload
-    async def create_file_attachment(
-        self, item_id: int, body: _models.FileAttachmentMultipartRequest, **kwargs: Any
-    ) -> None:
-        """create_file_attachment.
-
-        :param item_id: Required.
-        :type item_id: int
-        :param body: Required.
-        :type body: ~todo.models.FileAttachmentMultipartRequest
-        :return: None
-        :rtype: None
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-
-    @overload
-    async def create_file_attachment(self, item_id: int, body: JSON, **kwargs: Any) -> None:
-        """create_file_attachment.
-
-        :param item_id: Required.
-        :type item_id: int
-        :param body: Required.
-        :type body: JSON
-        :return: None
-        :rtype: None
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-
-    async def create_file_attachment(
-        self, item_id: int, body: Union[_models.FileAttachmentMultipartRequest, JSON], **kwargs: Any
-    ) -> None:
-        """create_file_attachment.
-
-        :param item_id: Required.
-        :type item_id: int
-        :param body: Is either a FileAttachmentMultipartRequest type or a JSON type. Required.
-        :type body: ~todo.models.FileAttachmentMultipartRequest or JSON
-        :return: None
-        :rtype: None
-        :raises ~corehttp.exceptions.HttpResponseError:
-        """
-        error_map: MutableMapping = {
-            304: ResourceNotModifiedError,
-        }
-        error_map.update(kwargs.pop("error_map", {}) or {})
-
-        _headers = kwargs.pop("headers", {}) or {}
-        _params = kwargs.pop("params", {}) or {}
-
-        cls: ClsType[None] = kwargs.pop("cls", None)
-
-        _body = body.as_dict() if isinstance(body, _model_base.Model) else body
-        _file_fields: List[str] = ["contents"]
-        _data_fields: List[str] = []
-        _files, _data = prepare_multipart_form_data(_body, _file_fields, _data_fields)
-
-        _request = build_todo_items_attachments_create_file_attachment_request(
-            item_id=item_id,
-            files=_files,
-            data=_data,
-            headers=_headers,
-            params=_params,
-        )
-        path_format_arguments = {
-            "endpoint": self._serialize.url("self._config.endpoint", self._config.endpoint, "str", skip_quote=True),
-        }
-        _request.url = self._client.format_url(_request.url, **path_format_arguments)
-
-        _stream = False
-        pipeline_response: PipelineResponse = await self._client.pipeline.run(_request, stream=_stream, **kwargs)
-
-        response = pipeline_response.http_response
-
-        if response.status_code not in [204]:
-            map_error(status_code=response.status_code, response=response, error_map=error_map)
-            error = None
-            if response.status_code == 404:
-                error = _failsafe_deserialize(_models.NotFoundErrorResponse, response.json())
-                raise ResourceNotFoundError(response=response, model=error)
-            elif 400 <= response.status_code <= 499:
-                error = _failsafe_deserialize(_models.Standard4XXResponse, response.json())
-            elif 500 <= response.status_code <= 599:
-                error = _failsafe_deserialize(_models.Standard5XXResponse, response.json())
+                error = _failsafe_deserialize(_models3.Standard5XXResponse, response.json())
             raise HttpResponseError(response=response, model=error)
 
         if cls:
