+++ todoApp/clients/python/todo/todoitems/models/__init__.py
@@ -7,16 +7,20 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from ._patch import *  # pylint: disable=unused-wildcard-import
 
-from ._operations import UsersOperations  # type: ignore
-from ._operations import TodoItemsOperations  # type: ignore
 
+from ._models import (  # type: ignore
+    InvalidTodoItem,
+    NotFoundErrorResponse,
+    TodoItemPatch,
+)
 from ._patch import __all__ as _patch_all
 from ._patch import *
 from ._patch import patch_sdk as _patch_sdk
 
 __all__ = [
-    "UsersOperations",
-    "TodoItemsOperations",
+    "InvalidTodoItem",
+    "NotFoundErrorResponse",
+    "TodoItemPatch",
 ]
 __all__.extend([p for p in _patch_all if p not in __all__])  # pyright: ignore
 _patch_sdk()
